Ongoing cleanup - separate code from builder.js
Create an input FSM for the builder
Separate the voxel world class
Create a WIP cell data class that'll be 'backend' data in a client-server infrastructure.
